requirements.txt
requests
azure-identity
azure-monitor-query
